pyproject.toml
[tool.poetry]
name = "crawling-project"
version = "0.1.0"
description = ""
authors = ["skwwnlpro <skwwnlpro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
webdriver-manager = "^4.0.1"
selenium = "^4.23.0"
pymysql = "^1.1.1"
yml = "^0.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
